Fix: ExperimentViewer renders marginals if diffMap empty
In #104 we refactored the diffMap out of search state into a new
diffSlice, and also defaulted it to {} (empty object) instead of
allowing it to be undefined / null.

This broke a check in ExperimentViewer on what to do when the diffMap
is empty.

Fix is to simply check the length of the diffMap rather than the
presence.