Allow retrieving progress updates from a search job started elsewhere
Instead of always adding a new job when opening a search, we can now
retrieve the job ID from a previous request; since results now come
through Redis streams and so are persistent, that means we can fetch the
results from the execution of the job, no matter when it started.

Right now, the streams stick around forever, which is not ideal. That's
related to the still-open question of how to balance information between
Postgres and Redis. It feels sensible in some sense to use Postgres for
long-term data and Redis for shorter-term, more realtime communication,
but situations like this don't necessarily map cleanly onto that
dichotomy.

Also, jobs as independent entities are presently a little bit redundant,
since we never create a new job when one already exists for a query. But
the idea is that we will want to be able to do things like rerun a job
for a query if something changes new agent runs are added, so that will
be more necessary later.